refactor(server): extract workspace remote access
