+++ scripts/max_lines_allowlist.txt
@@ -40,14 +40,14 @@ server/proliferate/server/cloud/runtime/bootstrap.py 922 server structure migrat
 server/proliferate/server/cloud/runtime/provision.py 1688 server structure migration debt for oversized domain-adjacent file
 server/proliferate/server/cloud/runtime_config/domain/resolver.py 611 server structure migration debt for oversized pure-domain file
 server/proliferate/server/cloud/runtime_config/service.py 1095 server structure migration debt for oversized service file
-server/proliferate/server/cloud/workspaces/models.py 890 server structure migration debt for oversized API models file
-server/proliferate/server/cloud/workspaces/service.py 1836 server structure migration debt for oversized service file
+server/proliferate/server/cloud/workspaces/models.py 861 server structure migration debt for oversized API models file
+server/proliferate/server/cloud/workspaces/service.py 1471 server structure migration debt for oversized service file
 server/tests/integration/schema_migration_assertions.py 878 existing server oversized test helper
 server/tests/integration/test_auth_flow.py 2057 existing server oversized integration test
 server/tests/integration/test_billing_accounting.py 871 existing server oversized integration test
 server/tests/integration/test_billing_api.py 2053 existing server oversized integration test
 server/tests/integration/test_cloud_agent_auth_api.py 1664 existing server oversized integration test
-server/tests/integration/test_cloud_api.py 2373 existing server oversized integration test
+server/tests/integration/test_cloud_api.py 2374 existing server oversized integration test
 server/tests/integration/test_cloud_commands_api.py 4890 existing server oversized integration test
 server/tests/integration/test_sandbox_profile_foundation.py 1224 existing server oversized integration test
 server/tests/integration/test_schema_migrations.py 925 existing server oversized integration test
