+++ scripts/server_boundaries_allowlist.txt
@@ -20,6 +20,7 @@ SERVICE_DB_ENGINE_IMPORT server/proliferate/server/cloud/commands/service.py 1 r
 SERVICE_DB_ENGINE_IMPORT server/proliferate/server/cloud/commands/transactions.py 1 Lane 7 explicit API-facing command commit boundary debt
 SERVICE_DB_ENGINE_IMPORT server/proliferate/server/cloud/live/service.py 1 remaining service-owned live event session boundary debt through session_ops
 SERVICE_DB_ENGINE_IMPORT server/proliferate/server/cloud/worker/transactions.py 1 Lane 7 explicit worker API pre-response commit boundary debt
+SERVICE_DB_ENGINE_IMPORT server/proliferate/server/cloud/workspaces/remote_access/service.py 1 remote-access connection flow preserves existing explicit session boundary debt through session_ops
 SERVICE_DB_ENGINE_IMPORT server/proliferate/server/cloud/workspaces/service.py 1 remaining service-owned workspace lifecycle session boundary debt through session_ops
 SERVICE_DB_ENGINE_IMPORT server/proliferate/server/cloud/workspaces/target_launch/service.py 1 target-launch orchestration preserves existing explicit command/session boundary debt through session_ops
 SERVICE_DB_METHOD_CALL server/proliferate/server/billing/service.py 34 remaining billing service-owned transaction/session boundary debt through session_ops
@@ -29,7 +30,8 @@ SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/commands/agent_auth_refre
 SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/commands/transactions.py 1 Lane 7 explicit API-facing command commit boundary debt
 SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/live/service.py 4 remaining live service-owned session/deferred boundary debt through session_ops
 SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/worker/transactions.py 2 Lane 7 explicit worker API pre-response commit boundary debt
-SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/workspaces/service.py 29 remaining workspace service-owned lifecycle transaction/session boundary debt through session_ops
+SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/workspaces/remote_access/service.py 4 remote-access connection flow preserves existing explicit session boundary debt through session_ops
+SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/workspaces/service.py 26 remaining workspace service-owned lifecycle transaction/session boundary debt through session_ops
 SERVICE_DB_METHOD_CALL server/proliferate/server/cloud/workspaces/target_launch/service.py 5 target-launch orchestration preserves existing explicit command/session boundary debt through session_ops
 DOMAIN_FORBIDDEN_IMPORT server/proliferate/server/cloud/agent_run_config/domain/resolve.py 1 transitional agent run config resolver imports product constants
 DOMAIN_FORBIDDEN_IMPORT server/proliferate/server/cloud/target_config/domain/policy.py 1 transitional target config policy imports product config during worker migration
