+++ server/proliferate/server/cloud/workspaces/models.py
@@ -11,7 +11,6 @@ from uuid import UUID
 from pydantic import BaseModel, Field
 
 from proliferate.constants.cloud import (
-    CloudAgentKind,
     CloudRuntimeEnvironmentStatus,
     CloudWorkspaceCleanupState,
     CloudWorkspaceStatus,
@@ -135,22 +134,6 @@ class UpdateCloudWorkspaceDisplayNameRequest(BaseModel):
     """
 
     display_name: str | None = Field(default=None, alias="displayName")
-
-
-class RemoteAccessRepoRef(BaseModel):
-    provider: str = "local"
-    owner: str = "local"
-    name: str
-    branch: str = "default"
-    base_branch: str | None = Field(default=None, alias="baseBranch")
-
-
-class BootstrapWorkspaceRemoteAccessRequest(BaseModel):
-    target_id: UUID = Field(alias="targetId")
-    anyharness_workspace_id: str = Field(alias="anyharnessWorkspaceId", min_length=1)
-    anyharness_session_id: str | None = Field(default=None, alias="anyharnessSessionId")
-    display_name: str | None = Field(default=None, alias="displayName")
-    repo: RemoteAccessRepoRef | None = None
 
 
 def _to_iso(value: object) -> str | None:
@@ -411,18 +394,6 @@ class WorkspaceRuntimeAuthState(BaseModel):
     last_applied_at: str | None = Field(
         default=None,
         serialization_alias="lastAppliedAt",
-    )
-
-
-class WorkspaceConnection(BaseModel):
-    runtime_url: str = Field(serialization_alias="runtimeUrl")
-    access_token: str = Field(serialization_alias="accessToken")
-    anyharness_workspace_id: str | None = Field(serialization_alias="anyharnessWorkspaceId")
-    runtime_generation: int = Field(serialization_alias="runtimeGeneration")
-    allowed_agent_kinds: list[CloudAgentKind] = Field(serialization_alias="allowedAgentKinds")
-    ready_agent_kinds: list[str] = Field(serialization_alias="readyAgentKinds")
-    runtime_auth: WorkspaceRuntimeAuthState = Field(
-        serialization_alias="runtimeAuth",
     )
